Create Take Profit Indicator v1.0 Released [Upper Version]
ere is the upper study of the Take Profit indicator.

When the TP line crosses below 0.1, it will generate an "Exit" bubble on your chart. If you prefer to change this value to something else, you are more than welcome to do so. You can read more about the trading rules here.

My suggestion: Use this in conjunction with the Buy the Dip indicator as a profit taker. Depending on your risk tolerance and trading style, you can also use it to short stocks.

Some Notes
There are two modes in this indicator: Exit mode and Short mode. Both are eventually the same, just different wording in the bubbles that you see on your chart. When Exit mode is enabled, it will say "Exit." When Short mode is enabled, it will say "Short."
The Meter label can be turned On or Off (by default it's set to On).