Create Take Profit Indicator v1.0 Released [Standard Version]
https://usethinkscript.com/threads/take-profit-indicator-v1-0-released-standard-version.3656/

Take Profit is similar to our Buy the Dip indicator. It's based on the Synthetic VIX by Larry Williams. The standard version is also an oscillator. This version uses his original formula rather than the inverse of it. Furthermore, I applied the PercentRank function (script by @diazlaz) to normalize the Synthetic VIX value. Jeff Swanson of EasyLanguageMastery proposed this normalization idea.

According to Swanson, we should go long when the oscillator is above 95% and short when it is below 5%. However, based on my observation, the 95% rule isn't as good as the original Buy the Dip indicator. But I'll leave that for you guys to decide. Furthermore, instead of using the 5% rule to short, I find it better when the oscillator is below 0.1 or even crosses below the Zero line (rare).

So the rule of this indicator is similar to BTD. Go short / take profit when TP (Take Profit) is at or below 0.1.

My suggestion: Use this in conjunction with the Buy the Dip indicator as a profit taker. Depending on your risk tolerance and trading style, you can also use it to short stocks.

Some notes
The Meter label can be turned On or Off via the indicator's settings (by default it's set to On)
The down arrows can be turned On or Off via the indicator's settings (by default it's set to Off)
The TP line changes color (from red to yellow) if it is at or below 0.1